argument tag: move name+argument parsing
- to tool_util.parser.util
- make ToolParameter.build work with InputSource
  instead of plain XML as commonly done in tests
- refactor stdio related function to tool_util.parser.stdio
  to avoid circular import
