+++ lib/galaxy/tools/parameters/basic.py
@@ -239,27 +239,21 @@ class ToolParameter(Dictifiable):
         return tool_dict
 
     @classmethod
-    def build(cls, tool, param):
+    def build(cls, tool, input_source):
         """Factory method to create parameter of correct type"""
-        param_name = cls.parse_name(param)
-        param_type = param.get('type')
+        input_source = ensure_input_source(input_source)
+        param_name = cls.parse_name(input_source)
+        param_type = input_source.get('type')
         if not param_type:
             raise ValueError("parameter '%s' requires a 'type'" % (param_name))
         elif param_type not in parameter_types:
             raise ValueError("parameter '%s' uses an unknown type '%s'" % (param_name, param_type))
         else:
-            return parameter_types[param_type](tool, param)
+            return parameter_types[param_type](tool, input_source)
 
     @staticmethod
     def parse_name(input_source):
-        name = input_source.get('name')
-        if name is None:
-            argument = input_source.get('argument')
-            if argument:
-                name = argument.lstrip('-').replace("-", "_")
-            else:
-                raise ValueError("parameter must specify a name.")
-        return name
+        return input_source.parse_name()
 
 
 class TextToolParameter(ToolParameter):
